requirements.txt
streamlit==1.43.0
sentence-transformers==3.0.1
transformers==4.44.2
huggingface-hub==0.24.6
# Pin torch to prevent massive re-downloads and resolution hangs
torch==2.2.2
python-dotenv==1.0.1
pydantic>=2.8.2
tenacity==8.3.0
# Pin these to prevent metadata/resolution slowness
pydantic-ai[vertexai]==0.0.24
psutil
more-itertools

pandas
PyMuPDF
pytesseract
Pillow
crawl4ai==0.3.73
requests
beautifulsoup4
pytest
psycopg2-binary
PyYAML
openai
starlette
uvicorn
google-cloud-bigquery>=3.10.0
google-cloud-aiplatform
google-auth>=2.20.0
protobuf==5.29.5
grpcio==1.76.0
grpcio-status==1.60.0
opentelemetry-proto==1.38.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
logfire==4.15.1
bcrypt
streamlit-authenticator
google-cloud-secret-manager
python-docx
google-cloud-storage
pdfplumber
